Refactor react components, cleanup
