+++ app/panels/Parameters/ParametersPanel.tsx
@@ -1,0 +1,12 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import styled from "styled-components";
+
+const ParametersPanel = styled.div`
+  display: flex;
+  flex-direction: column;
+`;
+
+export default ParametersPanel;
