+++ app/panels/Parameters/Scrollable.tsx
@@ -1,0 +1,14 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import styled from "styled-components";
+
+const Scrollable = styled.div`
+  height: 100%;
+  width: 100%;
+  overflow-y: auto;
+  position: absolute;
+`;
+
+export default Scrollable;
